Move block template functionality to block template provider
