+++ packages/js/product-editor/src/components/block-template-provider/index.ts
@@ -1,0 +1,1 @@
+export * from './block-template-provider';
