--- packages/js/product-editor/src/components/blocks-template/index.ts
@@ -1,1 +1,0 @@
-export * from './blocks-template';
